app/requirements.txt
streamlit-authenticator==0.1.4
Flask==2.1.2
Flask-WTF==1.0.1
WTForms==3.0.1
email-validator==1.2.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
